refactor: Use rollup to minify packages
