+++ lib/index.ts
@@ -1,7 +1,7 @@
 import { createHash } from 'crypto';
-import { IKeyStore, TKeyData, TKeyStoreInit, TSaveSettings } from '../lib/types';
-import { readFile, writeFile } from 'fs'
-;
+import { IKeyStore, TKeyData, TKeyStoreInit, TSaveSettings } from '../types/index.types';
+import { readFile, writeFile } from 'fs';
+
 export class KeyHolder implements IKeyStore {
 	/// ///////////////////////////////////////////
 	// Data and settings for the KeyStore class //
