+++ package.json
@@ -4,10 +4,10 @@
   "license": "MIT",
   "version": "0.6.1",
   "description": "Easy to use & lightweight management of key-value pairs 🚀✨",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
+  "main": "./dist/bundle/bundle.js",
+  "types": "./dist/types/index.types.d.ts",
   "scripts": {
-    "build": "tsc",
+    "build": "rimraf dist && tsc && rollup -c rollup.config.js",
     "jest": "jest",
     "lint": "eslint --ext .js,.ts ./lib",
     "prepare": "husky install",
@@ -26,6 +26,7 @@
   },
   "homepage": "https://github.com/BitsOfAByte/HoldMyKeys#readme",
   "devDependencies": {
+    "@rollup/plugin-typescript": "^8.3.1",
     "@types/jest": "^27.4.1",
     "@types/node": "^17.0.21",
     "@typescript-eslint/eslint-plugin": "^5.14.0",
@@ -33,6 +34,9 @@
     "eslint": "^8.11.0",
     "husky": "^7.0.0",
     "jest": "^27.5.1",
+    "rimraf": "^3.0.2",
+    "rollup": "^2.70.1",
+    "rollup-plugin-terser": "^7.0.2",
     "ts-jest": "^27.1.3",
     "typescript": "^4.6.2"
   }
